+++ src/Features/Core/Portable/Diagnostics/IDiagnosticAnalyzerService.cs
@@ -74,14 +74,14 @@ namespace Microsoft.CodeAnalysis.Diagnostics
         /// This API will only force complete analyzers that support span based analysis, i.e. compiler analyzer and
         /// <see cref="IBuiltInAnalyzer"/>s that support <see cref="DiagnosticAnalyzerCategory.SemanticSpanAnalysis"/>.
         /// For the rest of the analyzers, it will only return diagnostics if the analyzer has already been executed.
-        /// Use <see cref="GetDiagnosticsForSpanAsync(TextDocument, TextSpan?, Func{string, bool}?, bool, bool, CodeActionRequestPriority, Func{string, IDisposable?}?, DiagnosticKinds, CancellationToken)"/>
+        /// Use <see cref="GetDiagnosticsForSpanAsync(TextDocument, TextSpan?, Func{string, bool}?, bool, bool, CodeActionRequestPriority, Func{string, IDisposable?}?, DiagnosticKind, CancellationToken)"/>
         /// if you want to force complete all analyzers and get up-to-date diagnostics for all analyzers for the given span.
         /// </summary>
         Task<(ImmutableArray<DiagnosticData> diagnostics, bool upToDate)> TryGetDiagnosticsForSpanAsync(
             TextDocument document, TextSpan range, Func<string, bool>? shouldIncludeDiagnostic,
             bool includeSuppressedDiagnostics = false,
             CodeActionRequestPriority priority = CodeActionRequestPriority.None,
-            DiagnosticKinds diagnosticKinds = DiagnosticKinds.All,
+            DiagnosticKind diagnosticKinds = DiagnosticKind.All,
             CancellationToken cancellationToken = default);
 
         /// <summary>
@@ -97,7 +97,7 @@ namespace Microsoft.CodeAnalysis.Diagnostics
             bool includeCompilerDiagnostics, bool includeSuppressedDiagnostics = false,
             CodeActionRequestPriority priority = CodeActionRequestPriority.None,
             Func<string, IDisposable?>? addOperationScope = null,
-            DiagnosticKinds diagnosticKinds = DiagnosticKinds.All,
+            DiagnosticKind diagnosticKinds = DiagnosticKind.All,
             CancellationToken cancellationToken = default);
     }
 
@@ -106,7 +106,7 @@ namespace Microsoft.CodeAnalysis.Diagnostics
         public static Task<ImmutableArray<DiagnosticData>> GetDiagnosticsForSpanAsync(this IDiagnosticAnalyzerService service,
             TextDocument document, TextSpan range, string? diagnosticId = null,
             bool includeSuppressedDiagnostics = false, Func<string, IDisposable?>? addOperationScope = null,
-            DiagnosticKinds diagnosticKinds = DiagnosticKinds.All,
+            DiagnosticKind diagnosticKinds = DiagnosticKind.All,
             CancellationToken cancellationToken = default)
             => service.GetDiagnosticsForSpanAsync(document, range, diagnosticId, includeSuppressedDiagnostics, CodeActionRequestPriority.None, addOperationScope, diagnosticKinds, cancellationToken);
 
@@ -123,7 +123,7 @@ namespace Microsoft.CodeAnalysis.Diagnostics
             bool includeSuppressedDiagnostics = false,
             CodeActionRequestPriority priority = CodeActionRequestPriority.None,
             Func<string, IDisposable?>? addOperationScope = null,
-            DiagnosticKinds diagnosticKinds = DiagnosticKinds.All,
+            DiagnosticKind diagnosticKinds = DiagnosticKind.All,
             CancellationToken cancellationToken = default)
         {
             Func<string, bool>? shouldIncludeDiagnostic = diagnosticId != null ? id => id == diagnosticId : null;
